+++ packages/app/src/extend.ts
@@ -38,7 +38,11 @@ import { App } from './app'
 /**
  * Extends Request and Response objects with custom properties and methods
  */
-export const extendMiddleware = (app: App) => (req: Request, res: Response, next: NextFunction) => {
+export const extendMiddleware = <EngineOptions>(app: App) => (
+  req: Request,
+  res: Response<EngineOptions>,
+  next: NextFunction
+) => {
   const { settings } = app
 
   req.originalUrl = req.url
@@ -86,7 +90,7 @@ export const extendMiddleware = (app: App) => (req: Request, res: Response, next
   res.vary = setVaryHeader<Response>(res)
   res.cookie = setCookie<Request, Response>(req, res)
   res.clearCookie = clearCookie<Request, Response>(req, res)
-  res.render = renderTemplate(req, res, app)
+  res.render = renderTemplate<EngineOptions>(req, res, app)
   res.format = formatResponse(req, res, next)
   res.redirect = redirect(req, res, next)
   res.attachment = attachment<Response>(res)