+++ packages/cookie-parser/CHANGELOG.md
@@ -1,6 +1,14 @@
 # @tinyhttp/cookie-parser
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@1.0.12
+
 ## 1.0.11
+
 ### Patch Changes
 
 - regenerate changelogs