+++ packages/markdown/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @tinyhttp/markdown
 
+## 1.0.12
+
+### Patch Changes
+
+- Add template settings and caching for markdown
+- Updated dependencies [undefined]
+  - @tinyhttp/app@1.0.12
+
 ## 1.0.11
+
 ### Patch Changes
 
 - regenerate changelogs